[rsocket-cpp] Simplify warm-resumption integration test
